MAINT: Add nd versions for all other polynomial methods
